project/backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
Click==8.3.1
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.3.0
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv
